third_party/blink/web_tests/virtual/html-anchor-attribute-disabled/external/wpt/html/semantics/popovers/popover-top-layer-nesting-anchor.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Single popover=auto ancestor with dialog, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Single popover=auto ancestor with fullscreen, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Single popover=manual ancestor with dialog, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Single popover=manual ancestor with fullscreen, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Nested popover=auto ancestors with dialog, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Nested popover=auto ancestors with fullscreen, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Nested popover=auto ancestors, target is outer with dialog, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Nested popover=auto ancestors, target is outer with fullscreen, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Top layer inside of nested element with dialog, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
[FAIL] Top layer inside of nested element with fullscreen, anchor attribute
  assert_equals: Incorrect behavior expected true but got false
Harness: the test ran to completion.
